Configuração HOSTINGER (#5)
* :sparkles: FEAT: Star-roadmap Route

Adicionando a rota que inicia o roadmap pegando o id do usuário e o id do roadmap

* :sparkles: FEAT: Implementação do JWT e Bcrypt

* :boom: REMOVE: Delete outside domain

* :wrench: CHORE: Moving security directory

* :sparkles: FEAT: End-roamap route

* :construction:  FEATK: FAKER  TEST

* :construction: FAKER TEST 1.1

* :sparkles: FEAT: jwt

* :construction:  TEST: FAKER TEST 1.2

* 🧹CLEANUP: Renaming variables

Renomeando variáveis

* :books: DOCS: Crud Doc

Arquivo markdown explicando como funciona cada funçao nativa do Crud Interface

* construction sign TEST: FAKER TEST 1.3

* :construction: TEST: FAKER TEST 1.4

* :sparkles: FEAT: Faker

* FIX: Merge conflicts

* FIX: conflicts

* FIX: START/END RoadmapService

* 🧹 CLEANUP: Ident and remove comments

* :construction: TEST: UserAchievements

* FIX: Delete non-standard folder

* :recycle: REFACTOR: MapsStruct Refactor

* :sparkles: FEAT: TaskText Column Received JSONB

* Revert "Merge remote-tracking branch 'origin/Feat-Step' into backend"

This reverts commit 19a2b21a2458a0d66109997c85de1a5e21bed3a5, reversing
changes made to c2811aa64a190aa45144102c5e7bdae4b2227ae1.

* Revert ":sparkles: FEAT: TaskText Column Received JSONB"

This reverts commit c2811aa64a190aa45144102c5e7bdae4b2227ae1.

* FIX: TaskText Column Received JSONB

* :construction: TEST: StepRoute fix

* :construction:  TEST: StepRoute fix 1.1

* :construction: TEST: UserAchievement

* :construction: TEST: Mapstruct problem

* :construction: TEST: Mapstruct problem 1.1

* :construction: TEST: Mapstruct problem 1.2

* :construction: TEST: Mapstruct problem 1.3

* :wrench: CHORE: .gitignore

* :construction: TEST: Mapstruct problem 1.3.1

* :wrench: FIX: Mapstruct

* :sparkles: FEAT: Add Swagger

* Update compiler.xml

* FIX: Update pom.xml

* :wrench: FIX: .gitIgnore

* :sparkles: FEAT: User-Step/Lesson Routes

* :books: DOCS: Arquivo Payloads.md

* :books: DOCS: Fix

* FIX: changing file to another folder

* Update application.properties

* Update application.properties

* Update application.properties

* Update application.properties

---------

+++ src/main/java/com/thigas/quack/application/service/UserService.java
@@ -1,9 +1,8 @@
 package com.thigas.quack.application.service;
 
 import com.thigas.quack.adapter.dto.UserDTO;
-import com.thigas.quack.adapter.mapper.UserMapper;
-import com.thigas.quack.domain.entity.UserEntity;
 import com.thigas.quack.domain.repository.IUserRepository;
+import com.thigas.quack.infrastructure.persistence.entity.UserModel;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,40 +15,60 @@ import java.util.stream.StreamSupport;
 @Service
 public class UserService {
 
-    private final UserMapper userMapper = UserMapper.INSTANCE;
     @Autowired
     private IUserRepository userRepository;
 
+    @Autowired
+    private ObjectMapperService objectMapperService = new ObjectMapperService();
+
     public UserDTO create(UserDTO userDTO) {
-        UserEntity user = userMapper.dtoToEntity(userDTO);
-        UserEntity toSaveUser = userRepository.save(user);
-        return userMapper.entityToDto(toSaveUser);
+        UserModel user = objectMapperService.toModel(userDTO);
+        UserModel savedUser = userRepository.save(user);
+        return objectMapperService.toDto(savedUser);
     }
 
     public Optional<UserDTO> getById(int id) {
-        Optional<UserEntity> user = userRepository.findById(id);
-        return user.map(userMapper::entityToDto);
+        return userRepository.findById(id)
+                .map(objectMapperService::toDto);
     }
 
     public Iterable<UserDTO> getAll() {
-        Iterable<UserEntity> users = userRepository.findAll();
-        return StreamSupport.stream(users.spliterator(), false).map(userMapper::entityToDto)
+        Iterable<UserModel> users = userRepository.findAll();
+        return StreamSupport.stream(users.spliterator(), false)
+                .map(objectMapperService::toDto)
                 .collect(Collectors.toList());
     }
 
     public void update(UserDTO userDTO) {
-        UserEntity existingUser = userRepository.findById(userDTO.getId())
+        UserModel existingUser = userRepository.findById(userDTO.getId())
                 .orElseThrow(() -> new EntityNotFoundException("User not found"));
+
         if (userDTO.getName() != null) {
             existingUser.setName(userDTO.getName());
         }
         if (userDTO.getBornAt() != null) {
             existingUser.setBornAt(LocalDate.parse(userDTO.getBornAt()));
         }
+
         userRepository.save(existingUser);
     }
 
     public void delete(int id) {
+        if (!userRepository.existsById(id)) {
+            throw new EntityNotFoundException("User not found");
+        }
         userRepository.deleteById(id);
+    }
+
+    public Optional<UserDTO> findByEmail(String email) {
+        return userRepository.findByEmail(email)
+                .map(objectMapperService::toDto);
+    }
+
+
+    //TODO: Recuperar as outras classes, DTOS, Entities
+
+    public Boolean existsById(int userId) {
+        return userRepository.existsById(userId);
     }
 }
