+++ src/main/java/com/thigas/quack/application/service/UserTaskService.java
@@ -1,9 +1,10 @@
 package com.thigas.quack.application.service;
 
 import com.thigas.quack.adapter.dto.UserTaskDTO;
-import com.thigas.quack.adapter.mapper.UserTaskMapper;
 import com.thigas.quack.domain.entity.UserTaskEntity;
 import com.thigas.quack.domain.repository.IUserTaskRepository;
+import com.thigas.quack.infrastructure.persistence.entity.UserTaskModel;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,33 +15,48 @@ import java.util.stream.StreamSupport;
 @Service
 public class UserTaskService {
 
-    private final UserTaskMapper userTaskMapper = UserTaskMapper.INSTANCE;
     @Autowired
     private IUserTaskRepository userTaskRepository;
 
+    @Autowired
+    private ObjectMapperService objectMapperService;
+
     public UserTaskDTO create(UserTaskDTO userTaskDTO) {
-        UserTaskEntity userTask = userTaskMapper.dtoToEntity(userTaskDTO);
-        UserTaskEntity toSaveRoadmapUser = userTaskRepository.save(userTask);
-        return userTaskMapper.entityToDto(toSaveRoadmapUser);
+        UserTaskEntity userTaskEntity = objectMapperService.toEntity(userTaskDTO);
+        UserTaskModel savedUserTask = userTaskRepository.save(objectMapperService.toModel(userTaskEntity));
+        return objectMapperService.toDto(savedUserTask);
     }
+
 
     public Optional<UserTaskDTO> getById(int id) {
-        Optional<UserTaskEntity> userTask = userTaskRepository.findById(id);
-        return userTask.map(userTaskMapper::entityToDto);
+        return userTaskRepository.findById(id)
+                .map(objectMapperService::toDto);
     }
 
+
     public Iterable<UserTaskDTO> getAll() {
-        Iterable<UserTaskEntity> userTasks = userTaskRepository.findAll();
-        return StreamSupport.stream(userTasks.spliterator(), false).map(userTaskMapper::entityToDto)
+        Iterable<UserTaskModel> userTasks = userTaskRepository.findAll();
+        return StreamSupport.stream(userTasks.spliterator(), false)
+                .map(objectMapperService::toDto)
                 .collect(Collectors.toList());
     }
 
+
     public void update(UserTaskDTO userTaskDTO) {
-        UserTaskEntity userTask = userTaskMapper.dtoToEntity(userTaskDTO);
-        userTaskRepository.save(userTask);
+        UserTaskModel existingUserTask = userTaskRepository.findById(userTaskDTO.getId())
+                .orElseThrow(() -> new EntityNotFoundException("User task not found"));
+
+        if (userTaskDTO.getId() != 0) {
+            existingUserTask.setId(userTaskDTO.getId());
+        }
+
+        userTaskRepository.save(existingUserTask);
     }
 
     public void delete(int id) {
+        if (!userTaskRepository.existsById(id)) {
+            throw new EntityNotFoundException("User task not found");
+        }
         userTaskRepository.deleteById(id);
     }
 }
